+++ src/intTest/kotlin/com/github/gbrowser/TestTags.kt
@@ -1,0 +1,8 @@
+package com.github.gbrowser
+
+import org.junit.jupiter.api.Tag
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+@Tag("uitest")
+annotation class UITest