+++ src/main/kotlin/com/github/gbrowser/actions/GFavoriteWebAction.kt
@@ -1,0 +1,17 @@
+package com.github.gbrowser.actions
+
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.ui.jcef.JBCefBrowser
+import javax.swing.Icon
+
+
+class GFavoriteWebAction(private val webUrl: String, icon: Icon, private val jbCefBrowser: JBCefBrowser) :
+  AnAction({ webUrl }, icon) {
+
+  override fun actionPerformed(e: AnActionEvent) {
+    jbCefBrowser.cefBrowser.loadURL(webUrl)
+  }
+
+
+}