+++ src/main/kotlin/com/github/gbrowser/ui/toolwindow/base/toolwindowUtils.kt
@@ -1,0 +1,9 @@
+package com.github.gbrowser.ui.toolwindow.base
+
+import com.intellij.openapi.wm.ToolWindow
+import com.intellij.openapi.wm.ex.ToolWindowEx
+
+fun ToolWindow.dontHideOnEmptyContent() {
+  setToHideOnEmptyContent(false)
+  (this as? ToolWindowEx)?.emptyText?.text = ""
+}